+++ src/protocol/server.rs
@@ -12,16 +12,16 @@ use super::protocol::{CiphertextType, TagType};
 #[derive(Debug)]
 pub struct Server {
     clients_keys: Vec<(lb_vrf::keypair::PublicKey, lb_vrf::keypair::SecretKey)>,
-    kem_keys: HashMap<u8, (kem::PublicKey, kem::SecretKey)>,
-    comms: HashMap<u8, Vec<u8>>,
-    comms_server: HashMap<u8, Vec<u8>>,
-    opens_server: HashMap<u8, (Vec<u8>, Vec<u8>)>,
+    kem_keys: HashMap<u32, (kem::PublicKey, kem::SecretKey)>,
+    comms: HashMap<u32, Vec<u8>>,
+    comms_server: HashMap<u32, Vec<u8>>,
+    opens_server: HashMap<u32, (Vec<u8>, Vec<u8>)>,
     cis: Vec<Vec<u8>>,
     yis: Vec<Vec<u8>>,
     proofs: Vec<Proof>,
-    ns: HashMap<u8, Vec<u8>>,
-    k: HashMap<u8, Vec<u8>>,
-    ctxis: HashMap<u8, CiphertextType>,
+    ns: HashMap<u32, Vec<u8>>,
+    k: HashMap<u32, Vec<u8>>,
+    ctxis: HashMap<u32, CiphertextType>,
 }
 
 impl Default for Server {
@@ -47,29 +47,29 @@ impl Server {
         }
     }
 
-    pub fn receive_m1(&mut self, m1: (Vec<u8>, u8)) {
+    pub fn receive_m1(&mut self, m1: (Vec<u8>, u32)) {
         let (comm, id) = m1;
 
         self.add_commitment(comm, id);
     }
 
-    fn add_commitment(&mut self, comm: Vec<u8>, id: u8) {
+    fn add_commitment(&mut self, comm: Vec<u8>, id: u32) {
         self.comms.insert(id, comm);
     }
 
-    fn add_commitment_server(&mut self, comm: Vec<u8>, id: u8) {
+    fn add_commitment_server(&mut self, comm: Vec<u8>, id: u32) {
         self.comms_server.insert(id, comm);
     }
 
-    fn add_open_server(&mut self, open: (Vec<u8>, Vec<u8>), id: u8) {
+    fn add_open_server(&mut self, open: (Vec<u8>, Vec<u8>), id: u32) {
         self.opens_server.insert(id, open);
     }
 
-    pub fn get_kem_keypair(&self, index: u8) -> (kem::PublicKey, kem::SecretKey) {
+    pub fn get_kem_keypair(&self, index: u32) -> (kem::PublicKey, kem::SecretKey) {
         self.kem_keys.get(&index).unwrap().clone()
     }
 
-    pub fn set_kem_keypair(&mut self, keys: (kem::PublicKey, kem::SecretKey), index: u8) {
+    pub fn set_kem_keypair(&mut self, keys: (kem::PublicKey, kem::SecretKey), index: u32) {
         self.kem_keys.insert(index, keys);
     }
 
@@ -83,42 +83,42 @@ impl Server {
         self.clients_keys.clone()
     }
 
-    pub fn get_ctxis(&self) -> HashMap<u8, (Ciphertext, Vec<u8>, TagType)> {
+    pub fn get_ctxis(&self) -> HashMap<u32, (Ciphertext, Vec<u8>, TagType)> {
         self.ctxis.clone()
     }
 
-    pub fn get_comms(&self) -> HashMap<u8, Vec<u8>> {
+    pub fn get_comms(&self) -> HashMap<u32, Vec<u8>> {
         self.comms.clone()
     }
 
-    pub fn get_comms_server(&self) -> HashMap<u8, Vec<u8>> {
+    pub fn get_comms_server(&self) -> HashMap<u32, Vec<u8>> {
         self.comms_server.clone()
     }
 
-    pub fn get_opens_server(&self) -> HashMap<u8, (Vec<u8>, Vec<u8>)> {
+    pub fn get_opens_server(&self) -> HashMap<u32, (Vec<u8>, Vec<u8>)> {
         self.opens_server.clone()
     }
 
-    pub fn set_ns(&mut self, index: u8, ns: Vec<u8>) {
+    pub fn set_ns(&mut self, index: u32, ns: Vec<u8>) {
         self.ns.insert(index, ns);
     }
 
-    fn set_ctxi(&mut self, ctxi: CiphertextType, id: u8) {
+    fn set_ctxi(&mut self, ctxi: CiphertextType, id: u32) {
         self.ctxis.insert(id, ctxi);
     }
 
-    pub fn set_k(&mut self, key: Vec<u8>, index: u8) {
+    pub fn set_k(&mut self, key: Vec<u8>, index: u32) {
         let mut hasher = Sha3_256::new();
         hasher.update(key);
         let hashed_k: Vec<u8> = hasher.finalize().to_vec();
         self.k.insert(index, hashed_k);
     }
 
-    pub fn get_key(&mut self, index: u8) -> Vec<u8> {
+    pub fn get_key(&mut self, index: u32) -> Vec<u8> {
         self.k.get(&index).unwrap().to_vec()
     }
 
-    pub fn get_ns(&self, index: u8) -> Vec<u8> {
+    pub fn get_ns(&self, index: u32) -> Vec<u8> {
         self.ns.get(&index).unwrap().clone()
     }
 
@@ -145,7 +145,7 @@ impl Server {
         client.receive_m2(m2);
     }
 
-    pub fn receive_m3(&mut self, m3: (Vec<u8>, u8)) {
+    pub fn receive_m3(&mut self, m3: (Vec<u8>, u32)) {
         let (comm_s, id) = m3;
         self.add_commitment_server(comm_s, id);
     }
@@ -154,7 +154,7 @@ impl Server {
         client.receive_m4(m4);
     }
 
-    pub fn receive_m5(&mut self, m5: (CiphertextType, (Vec<u8>, Vec<u8>), u8)) {
+    pub fn receive_m5(&mut self, m5: (CiphertextType, (Vec<u8>, Vec<u8>), u32)) {
         let (ctxi, open_s, id) = m5;
         self.add_open_server(open_s, id);
         self.set_ctxi(ctxi, id);
